+++ sdk/src/client/builder.rs
@@ -247,6 +247,12 @@ impl ClientBuilder {
         self
     }
 
+    /// Set maximum parallel API requests.
+    pub fn with_max_parallel_api_requests(mut self, max_parallel_api_requests: usize) -> Self {
+        self.max_parallel_api_requests = max_parallel_api_requests;
+        self
+    }
+
     /// Build the Client instance.
     #[cfg(not(target_family = "wasm"))]
     pub async fn finish(self) -> Result<Client> {